requirements.txt
arrow==0.14.7
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-MySQLdb==1.0.1
ics==0.7
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
python-dateutil==2.8.2
six==1.16.0
TatSu==5.8.0
Werkzeug==2.0.3
